refactor: RecordService의 서비스를 CompetitionService와 ParticipantService에 통합
